Clean up some residual confusion between OIDs and RelFileNumbers.
Commit c609f59297bd00e17af4db5dfdd5c8701378452b missed a few places
where we are referring to the number used as a part of the relation
filename as an "OID". We now want to call that a "RelFileNumber".

Some of these places actually made it sound like the OID in question
is pg_class.oid rather than pg_class.relfilenode, which is especially
good to clean up.

Dilip Kumar with some editing by me.